+++ api/pkg/statuses/db/module.go
@@ -1,0 +1,7 @@
+package db
+
+import "getsturdy.com/api/pkg/di"
+
+func Module(c *di.Container) {
+	c.Register(New)
+}
