+++ api/pkg/statuses/enterprise/module.go
@@ -1,0 +1,10 @@
+package enterprise
+
+import (
+	"getsturdy.com/api/pkg/di"
+	"getsturdy.com/api/pkg/statuses/enterprise/graphql"
+)
+
+func Module(c *di.Container) {
+	c.Import(graphql.Module)
+}
