+++ api/pkg/statuses/service/service.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"fmt"
 
+	"getsturdy.com/api/pkg/events"
 	"getsturdy.com/api/pkg/statuses"
 	db_statuses "getsturdy.com/api/pkg/statuses/db"
-	"getsturdy.com/api/pkg/events"
 
 	"go.uber.org/zap"
 )
